refactor: restructure package
